Rename lalr package to lookahead
