--- parser/lr/lalr/example_test.go
@@ -1,1 +1,0 @@
-package lalr_test
