+++ parser/lr/lookahead/example_test.go
@@ -1,0 +1,1 @@
+package lookahead_test
