+++ parser/lr/lookahead/lookahead.go
@@ -1,4 +1,4 @@
-// Package lalr provides data structures and algorithms for building Look-Ahead LR (LALR) parsers.
+// Package lookahead provides data structures and algorithms for building Look-Ahead LR (LALR) parsers.
 // An LALR parser is a bottom-up parser for the class of LR(1) grammars.
 //
 // An LALR parser, similar to SLR, uses the canonical LR(0) items to construct the state machine (DFA),
@@ -6,9 +6,9 @@
 // LALR merges states with identical core LR(0) items but handles lookahead symbols for each merged state separately,
 // making it more precise than SLR and avoids many conflicts that SLR might encounter.
 // LALR is more powerful than SLR as it can handle a wider range of grammars, including most programming languages.
-// However, it is less powerful than canonical LR(1) because state merging
-// can lose distinctions in lookahead contexts, potentially leading to conflicts for some grammars.
+// However, it is less powerful than canonical LR because state merging can lose distinctions in lookahead contexts,
+// potentially leading to conflicts for some grammars.
 //
 // For more details on parsing theory,
 // refer to "Compilers: Principles, Techniques, and Tools (2nd Edition)".
-package lalr
+package lookahead
