+++ parser/lr/lookahead/lookahead_test.go
@@ -1,0 +1,1 @@
+package lookahead
